bugfixing: resetChecksumChunk() causes deadlock
REASON: resetChecksumChunk() method will reset the write buffer in
DFSOutputStream. It requires lock on DFSOutputStream object. In the
previous design, we use ChecksumChunk() to adjust the buffer size when
DataStreamer detects seek event. Note that DataStreamer is an
independent thread. A dead lock can occur when the main thread flush
some data written after seek. The data amount can be too small to
trigger the DataStreamer since the seek. Therefore, beholding the lock
of DFSOutputStream object, the main thread triggered the DataStreamer
now. And only till now does the DataStreamer found it has to seek to a
new position to write the data. On performing the seek in
endblock(newpos), the DataStreamer will use resetChecksumChunk() to
adjust the write buffer. There rise two problems: 1) it will never get
the lock on DFSOutputStream object since the main thread holding it IS
waiting for the DataStreamer thread to finish its job; 2) the data is
written to a buffer with wrong size--resetChecksumChunk() should be done
before any new data is written since last seeking. Here is the new
design: in DFSOutputStream.seek() reset the chunk size and write buffer
before immediately. Note tested yet.