skip checksum at loading.
In BlockReceiver.computePartialChunkCrc(): We do not compare it so far:
for appending only scenario, that is all data we get. But for overwrite,
we may need to write in the middle of a chunk. A 'perfect' way is to
read till the end of the chunk and compare. But since we are targeting
memory based file system. We just skip it for now.